Move Pulse consumers into the Pulse service
