+++ treeherder/etl/management/commands/read_pulse_pushes.py
@@ -3,8 +3,8 @@ from django.core.exceptions import ImproperlyConfigured
 from django.core.management.base import BaseCommand
 from kombu import Exchange
 
-from treeherder.etl.pulse_consumer import PushConsumer
-from treeherder.services.pulse import pulse_conn
+from treeherder.services.pulse import (PushConsumer,
+                                       pulse_conn)
 
 
 class Command(BaseCommand):
