+++ treeherder/services/pulse/__init__.py
@@ -1,5 +1,9 @@
 from .connection import pulse_conn
+from .consumers import (JobConsumer,
+                        PushConsumer)
 
 __all__ = [
+    "JobConsumer",
+    "PushConsumer",
     "pulse_conn",
 ]
